- adding a new template for the load screen - ripping out all of the static temp data classes, and replacing their references with the data loaded from the Database - a giant bundle of data layer code

+++ Assets/Scripts/Helpers/Battle/Actions/ShootAction.cs
@@ -97,7 +97,7 @@ namespace Iam.Scripts.Helpers.Battle.Actions
             int roll = RNG.GetIntBelowMax(0, soldier.Soldier.Body.TotalProbabilityMap[soldier.Stance]);
             foreach (HitLocation location in soldier.Soldier.Body.HitLocations)
             {
-                int locationChance = location.Template.HitProbabilityMap[soldier.Stance];
+                int locationChance = location.Template.HitProbabilityMap[(int)soldier.Stance];
                 if (roll < locationChance)
                 {
                     return location;
